refactor: split duplicate finder app logic
