+++ src/app/duplicates.rs
@@ -1,0 +1,17 @@
+mod actions;
+mod input;
+mod model;
+mod scan;
+
+#[cfg(test)]
+mod tests;
+
+use super::*;
+use crate::app::jobs::DuplicateScanRequest;
+use anyhow::Result;
+use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
+use std::{
+    collections::HashSet,
+    fs,
+    path::{Path, PathBuf},
+};
